Ultima version

+++ megazarl-presentacion/src/main/java/edu/student/itson/dissof/megazarl/presentacion/ProductosVenta.java
@@ -299,7 +299,7 @@ public class ProductosVenta extends JFrame implements IProductosVenta, IVista {
                             COLOR_FONDO_SEPARADOR,
                             MARGEN_VERTICAL_COMPONENTES,
                             MARGEN_HORIZONTAL_COMPONENTES, 
-                            (Long) informacionProductoInicio.get("Id"),
+                            informacionProductoInicio.get("Id"),
                             (String) informacionProductoInicio.get("DireccionImagenProducto"),
                             (String) informacionProductoInicio.get("Nombre"), 
                             (String) informacionProductoInicio.get("Variedad"),
